requirements.txt
opencv-python
numpy
onnxruntime
torch
torchvision
Pillow
mediapipe
